+++ Project1/include/BeagleBone/tmp102_sensor.h
@@ -41,6 +41,6 @@ typedef enum temperature_unit
 #define TMP102_getTemp_Kelvin(p_temp)   TMP102_getTemp(p_temp, KELVIN)
 #define TMP102_getTemp_Fahren(p_temp)   TMP102_getTemp(p_temp, FAHREN)
 
-int TMP102_getTemp(double *temp, TEMPERATURE_UNIT_T unit);
+int TMP102_getTemp(float *temp, TEMPERATURE_UNIT_T unit);
 
 #endif
